Refactor common code in gh-dxp
